继续完善

+++ WZMediaPlayer/Player/Player/player_impl/AudioDecoder.cpp
@@ -186,7 +186,7 @@ void AudioDecoder::getVolume(float* volume){
     *volume = mVolume;
 }
 
-uint32_t* AudioDecoder::getAudioClockPtr(){
+int64_t* AudioDecoder::getAudioClockPtr(){
     return &mCurrentAudioClock;
 }
